Cambio tipos de dato para precio y peso para que permitan decimales

+++ src/main/java/StrategyEnvio.java
@@ -1,4 +1,4 @@
 public abstract class StrategyEnvio {
-    abstract int calcularPrecio(boolean largaDistancia,int kiloPaquete);
+    abstract double calcularPrecio(boolean largaDistancia,double kiloPaquete);
 }
